chore: move review fetching
