+++ package.json
@@ -40,9 +40,11 @@
     "@emotion/react": "^11.4.1",
     "@emotion/styled": "^11.3.0",
     "@react-icons/all-files": "^4.1.0",
+    "@shopify/shopify-api": "^2.0.0",
     "@snek-at/jaen": "^2.0.0",
     "@snek-at/jaen-pages": "^2.1.0",
     "electron": "^14.0.0",
+    "esm": "^3.2.25",
     "framer-motion": "^4.1.17",
     "gatsby": "^3.12.1",
     "gatsby-plugin-gatsby-cloud": "^2.11.0",
@@ -54,6 +56,7 @@
     "gatsby-source-filesystem": "^3.13.0",
     "gatsby-source-shopify": "^6.5.1",
     "gatsby-transformer-sharp": "^3.13.0",
+    "ngrok": "^4.3.0",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
